requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.3
django-allauth==0.35.0
django-debug-toolbar==1.9.1
django-webpack-loader==0.6.0
idna==2.6
oauthlib==2.0.6
pkg-resources==0.0.0
psycopg2==2.7.4
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
sqlparse==0.2.4
urllib3==1.22
